Authentication: Fix signature issue when running locally
If you run the emulator via your IDE or `dotnet run`, the signature
validation fails because the "host" includes a port (i.e.
`localhost:5000`) on the emulator side but it does not include a port on
the client side (since via the client SDK, Uri.Host does not return the
port component). This causes the signature validation to fail.

This strips the port component from the Host header if it is present, as
well as it breaks apart the several operations involved in computing a
signature into separate variables to aid in debugging.