Update Postman instructions to support local debugging
This replaces `pm.request.url.getRemote()` with `pm.request.url.getHost()` in the Postman pre-request script. This matches the behavior fix in my other PR #38 to support local debugging, as the signed host header is sent as just `localhost` (rather than with the port) by the SDK. 

This change has been tested to be backwards compatible with App Config running on a standard port, including live Azure App Config.