Add docker-compose file with better local debugging support
This adds a docker-compose.yml file and updates the README to simplify
running the app to just `docker-compose up -d`. This should greatly
simplify the experience for new users running the emulator. The compose
file also uses named volumes for easy persistence and access from the
host for the keypair and database files.

In order to better support local debugging without having to have access
to system folders, this uses an environment variable when running in
Docker for the database and cert/key files, but defaults to a "data"
folder within the repo so that the IDE/CLI does not need special
permissions to /var (on macOS/Linux).

Additionally, the openssl command has been moved into `keygen.sh` which
can be run locally for local development, but is also run by the
container and is aware of those Docker Compose environment variables,
falling back to the local data folder.